UUID fix
Removed kubwa UUID lookup due to the fact that it no longer
exists. Now checks w-hat, and falls back to SL Search.

This should find most keys, but not those who's profile is
hidden from search.

Issue: Kubwa returned a correctly capitalized username along
with the key, but neither w-hat or search do that. For now,
whatever you type in will be your capitalization. This could be
fixed by requesting it from llRequestUsername in-world.